✨ 게시물 추가 다이얼로그 widget으로 변경
