+++ src/widgets/addPostDialog/ui/index.ts
@@ -1,0 +1,1 @@
+export * from './AddPostDialog';
